Refactor Connection Schemas
