+++ equinix/fabric_cloud_router_read_schema.go
@@ -1,6 +1,9 @@
 package equinix
 
-import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+import (
+	equinix_schema "github.com/equinix/terraform-provider-equinix/internal/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
 
 func readPackageSch() map[string]*schema.Schema {
 	return map[string]*schema.Schema{
@@ -72,7 +75,7 @@ func readCloudRouterResourceSchema() map[string]*schema.Schema {
 			Computed:    true,
 			Description: "Captures Fabric Cloud Router lifecycle change information",
 			Elem: &schema.Resource{
-				Schema: readChangeLogSch(),
+				Schema: equinix_schema.ChangeLogSch(),
 			},
 		},
 		"type": {
@@ -85,7 +88,7 @@ func readCloudRouterResourceSchema() map[string]*schema.Schema {
 			Computed:    true,
 			Description: "Fabric Cloud Router location",
 			Elem: &schema.Resource{
-				Schema: readLocationSch(),
+				Schema: equinix_schema.LocationSch(),
 			},
 		},
 		"project": {
@@ -102,7 +105,7 @@ func readCloudRouterResourceSchema() map[string]*schema.Schema {
 			Computed:    true,
 			Description: "Customer account information that is associated with this Fabric Cloud Router",
 			Elem: &schema.Resource{
-				Schema: readAccountSch(),
+				Schema: equinix_schema.AccountSch(),
 			},
 		},
 		"order": {
@@ -110,7 +113,7 @@ func readCloudRouterResourceSchema() map[string]*schema.Schema {
 			Computed:    true,
 			Description: "Order information related to this Fabric Cloud Router",
 			Elem: &schema.Resource{
-				Schema: readOrderSch(),
+				Schema: equinix_schema.OrderSch(),
 			},
 		},
 		"notifications": {
@@ -118,7 +121,7 @@ func readCloudRouterResourceSchema() map[string]*schema.Schema {
 			Computed:    true,
 			Description: "Preferences for notifications on Fabric Cloud Router configuration or status changes",
 			Elem: &schema.Resource{
-				Schema: readNotificationSch(),
+				Schema: equinix_schema.NotificationSch(),
 			},
 		},
 	}
