--- equinix/fabric_connection_schema.go
@@ -1,901 +1,0 @@
-package equinix
-
-import (
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
-)
-
-var createServiceTokenRes = &schema.Resource{
-	Schema: createServiceTokenSch(),
-}
-
-func createServiceTokenSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"type": {
-			Type:         schema.TypeString,
-			Optional:     true,
-			ValidateFunc: validation.StringInSlice([]string{"VC_TOKEN"}, true),
-			Description:  "Token type - VC_TOKEN",
-		},
-		"href": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "An absolute URL that is the subject of the link's context",
-		},
-		"uuid": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Equinix-assigned service token identifier",
-		},
-		"description": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Service token description",
-		},
-	}
-}
-
-var createLocationRes = &schema.Resource{
-	Schema: createLocationSch(),
-}
-
-func createLocationSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"region": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Computed:    true,
-			Description: "Access point region",
-		},
-		"metro_name": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Computed:    true,
-			Description: "Access point metro name",
-		},
-		"metro_code": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Computed:    true,
-			Description: "Access point metro code",
-		},
-		"ibx": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Computed:    true,
-			Description: "IBX Code",
-		},
-	}
-}
-
-func createVirtualGatewaySch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"uuid": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Equinix-assigned virtual gateway identifier",
-		},
-		"href": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Unique Resource Identifier",
-		},
-	}
-}
-
-func createNetworkSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"uuid": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Equinix-assigned Network identifier",
-		},
-		"href": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Unique Resource Identifier",
-		},
-	}
-}
-
-var createGatewayProjectSchRes = &schema.Resource{
-	Schema: createGatewayProjectSch(),
-}
-
-func createGatewayProjectSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"project_id": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Optional:    true,
-			Description: "Project Id",
-		},
-		"href": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Unique Resource URL",
-		},
-	}
-}
-
-var createServiceProfileSchRes = &schema.Resource{
-	Schema: createServiceProfileSch(),
-}
-
-func createServiceProfileSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"href": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Service Profile URI response attribute",
-		},
-		"type": {
-			Type:         schema.TypeString,
-			Required:     true,
-			ValidateFunc: validation.StringInSlice([]string{"L2_PROFILE", "L3_PROFILE", "ECIA_PROFILE", "ECMC_PROFILE"}, true),
-			Description:  "Service profile type - L2_PROFILE, L3_PROFILE, ECIA_PROFILE, ECMC_PROFILE",
-		},
-		"name": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Customer-assigned service profile name",
-		},
-		"uuid": {
-			Type:        schema.TypeString,
-			Required:    true,
-			Description: "Equinix assigned service profile identifier",
-		},
-		"description": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "User-provided service description",
-		},
-		"access_point_type_configs": {
-			Type:        schema.TypeList,
-			Computed:    true,
-			Description: "Access point config information",
-			Elem: &schema.Resource{
-				Schema: createAccessPointTypeConfigSch(),
-			},
-		},
-	}
-}
-
-func createAccessPointTypeConfigSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"type": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Type of access point type config - VD, COLO",
-		},
-		"uuid": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Equinix-assigned access point type config identifier",
-		},
-	}
-}
-
-var createAccessPointLinkProtocolSchRes = &schema.Resource{
-	Schema: createAccessPointLinkProtocolSch(),
-}
-
-func createAccessPointLinkProtocolSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"type": {
-			Type:         schema.TypeString,
-			Optional:     true,
-			Description:  "Type of the link protocol - UNTAGGED, DOT1Q, QINQ, EVPN_VXLAN",
-			ValidateFunc: validation.StringInSlice([]string{"UNTAGGED", "DOT1Q", "QINQ", "EVPN_VXLAN"}, true),
-		},
-		"vlan_tag": {
-			Type:        schema.TypeInt,
-			Optional:    true,
-			Computed:    true,
-			Description: "Vlan Tag information, vlanTag value specified for DOT1Q connections",
-		},
-		"vlan_s_tag": {
-			Type:        schema.TypeInt,
-			Optional:    true,
-			Computed:    true,
-			Description: "Vlan Provider Tag information, vlanSTag value specified for QINQ connections",
-		},
-		"vlan_c_tag": {
-			Type:        schema.TypeInt,
-			Optional:    true,
-			Computed:    true,
-			Description: "Vlan Customer Tag information, vlanCTag value specified for QINQ connections",
-		},
-	}
-}
-
-var createAccessPointVirtualDeviceRes = &schema.Resource{
-	Schema: createAccessPointVirtualDeviceSch(),
-}
-
-func createAccessPointVirtualDeviceSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"href": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Unique Resource Identifier",
-		},
-		"uuid": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Equinix-assigned Virtual Device identifier",
-		},
-		"type": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Virtual Device type",
-		},
-		"name": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Customer-assigned Virtual Device Name",
-		},
-	}
-}
-
-func createAccessPointInterface() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"uuid": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Equinix-assigned interface identifier",
-		},
-		"id": {
-			Type:        schema.TypeInt,
-			Computed:    true,
-			Optional:    true,
-			Description: "id",
-		},
-		"type": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Interface type",
-		},
-	}
-}
-
-func createFabricConnectionRoutingProtocol() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"type": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Routing Protocol type",
-		},
-		"uuid": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Equinix-assigned Routing protocol identifier",
-		},
-		"state": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Routing protocol instance state",
-		},
-	}
-}
-
-var createPortRes = &schema.Resource{
-	Schema: createPortSch(),
-}
-
-func createPortSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"uuid": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Computed:    true,
-			Description: "Equinix-assigned Port identifier",
-		},
-		"href": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Unique Resource Identifier",
-		},
-		"name": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Port name",
-		},
-		"redundancy": {
-			Type:        schema.TypeSet,
-			Computed:    true,
-			Description: "Redundancy Information",
-			Elem: &schema.Resource{
-				Schema: createPortRedundancySch(),
-			},
-		},
-	}
-}
-
-func createConnectionSideAccessPointRes() *schema.Resource {
-	return &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"type": {
-				Type:         schema.TypeString,
-				Optional:     true,
-				ValidateFunc: validation.StringInSlice([]string{"COLO", "VD", "VG", "SP", "IGW", "SUBNET", "CLOUD_ROUTER", "NETWORK"}, true),
-				Description:  "Access point type - COLO, VD, VG, SP, IGW, SUBNET, CLOUD_ROUTER, NETWORK",
-			},
-			"authentication_key": {
-				Type:        schema.TypeString,
-				Optional:    true,
-				Description: "Authentication key for provider based connections",
-			},
-			"account": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Computed:    true,
-				Description: "Account",
-				Elem: &schema.Resource{
-					Schema: createAccountSch(),
-				},
-			},
-			"location": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Computed:    true,
-				Description: "Access point location",
-				MaxItems:    1,
-				Elem: &schema.Resource{
-					Schema: createLocationSch(),
-				},
-			},
-			"port": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Description: "Port access point information",
-				MaxItems:    1,
-				Elem: &schema.Resource{
-					Schema: createPortSch(),
-				},
-			},
-			"profile": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Description: "Service Profile",
-				MaxItems:    1,
-				Elem: &schema.Resource{
-					Schema: createServiceProfileSch(),
-				},
-			},
-			"gateway": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Deprecated:  "use router attribute instead; gateway is no longer a part of the supported backend",
-				Description: "Cloud Router access point information",
-				MaxItems:    1,
-				Elem: &schema.Resource{
-					Schema: createVirtualGatewaySch(),
-				},
-			},
-			"router": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Description: "Cloud Router access point information",
-				MaxItems:    1,
-				Elem: &schema.Resource{
-					Schema: createVirtualGatewaySch(),
-				},
-			},
-			"network": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Description: "network access point information",
-				MaxItems:    1,
-				Elem: &schema.Resource{
-					Schema: createNetworkSch(),
-				},
-			},
-			"link_protocol": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Description: "Connection link protocol",
-				MaxItems:    1,
-				Elem: &schema.Resource{
-					Schema: createAccessPointLinkProtocolSch(),
-				},
-			},
-			"virtual_device": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Description: "Virtual device",
-				MaxItems:    1,
-				Elem:        &schema.Resource{Schema: createAccessPointVirtualDeviceSch()},
-			},
-			"interface": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Description: "Virtual device interface",
-				MaxItems:    1,
-				Elem: &schema.Resource{
-					Schema: createAccessPointInterface(),
-				},
-			},
-			"seller_region": {
-				Type:        schema.TypeString,
-				Optional:    true,
-				Description: "Access point seller region",
-			},
-			"peering_type": {
-				Type:         schema.TypeString,
-				Optional:     true,
-				ValidateFunc: validation.StringInSlice([]string{"PRIVATE", "MICROSOFT", "PUBLIC", "MANUAL"}, true),
-				Description:  "Peering Type- PRIVATE,MICROSOFT,PUBLIC, MANUAL",
-			},
-			"routing_protocols": {
-				Type:        schema.TypeList,
-				Optional:    true,
-				Description: "Access point routing protocols configuration",
-				Elem: &schema.Resource{
-					Schema: createFabricConnectionRoutingProtocol(),
-				},
-			},
-			"provider_connection_id": {
-				Type:        schema.TypeString,
-				Optional:    true,
-				Description: "Provider assigned Connection Id",
-			},
-		},
-	}
-}
-
-func createFabricConnectionSideRes() *schema.Resource {
-	return &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"service_token": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Description: "For service token based connections, Service tokens authorize users to access protected resources and services. Resource owners can distribute the tokens to trusted partners and vendors, allowing selected third parties to work directly with Equinix network assets",
-				MaxItems:    1,
-				Elem: &schema.Resource{
-					Schema: createServiceTokenSch(),
-				},
-			},
-			"access_point": {
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Description: "Point of access details",
-				MaxItems:    1,
-				Elem:        createConnectionSideAccessPointRes(),
-			},
-			"additional_info": {
-				Type:        schema.TypeList,
-				Optional:    true,
-				Description: "Connection side additional information",
-				Elem: &schema.Resource{
-					Schema: createAdditionalInfoSch(),
-				},
-			},
-		},
-	}
-}
-
-var createRedundancyRes = &schema.Resource{
-	Schema: createRedundancySch(),
-}
-
-func createRedundancySch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"group": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Optional:    true,
-			Description: "Redundancy group identifier",
-		},
-		"priority": {
-			Type:         schema.TypeString,
-			Optional:     true,
-			ValidateFunc: validation.StringInSlice([]string{"PRIMARY", "SECONDARY"}, true),
-			Description:  "Priority type- PRIMARY, SECONDARY",
-		},
-	}
-}
-
-var createPortRedundancyRes = &schema.Resource{
-	Schema: createPortRedundancySch(),
-}
-
-func createPortRedundancySch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"priority": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Priority type- PRIMARY, SECONDARY",
-		},
-	}
-}
-
-var createChangeLogRes = &schema.Resource{
-	Schema: createChangeLogSch(),
-}
-
-func createChangeLogSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"created_by": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Created by User Key",
-		},
-		"created_by_full_name": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Created by User Full Name",
-		},
-		"created_by_email": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Created by User Email Address",
-		},
-		"created_date_time": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Created by Date and Time",
-		},
-		"updated_by": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Updated by User Key",
-		},
-		"updated_by_full_name": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Updated by User Full Name",
-		},
-		"updated_by_email": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Updated by User Email Address",
-		},
-		"updated_date_time": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Updated by Date and Time",
-		},
-		"deleted_by": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Deleted by User Key",
-		},
-		"deleted_by_full_name": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Deleted by User Full Name",
-		},
-		"deleted_by_email": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Deleted by User Email Address",
-		},
-		"deleted_date_time": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Deleted by Date and Time",
-		},
-	}
-}
-
-var createOrderRes = &schema.Resource{
-	Schema: createOrderSch(),
-}
-
-func createOrderSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"purchase_order_number": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Purchase order number",
-		},
-		"billing_tier": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Billing tier for connection bandwidth",
-		},
-		"order_id": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Order Identification",
-		},
-		"order_number": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Order Reference Number",
-		},
-	}
-}
-
-var createAccountRes = &schema.Resource{
-	Schema: createAccountSch(),
-}
-
-func createAccountSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"account_number": {
-			Type:        schema.TypeInt,
-			Computed:    true,
-			Description: "Account Number",
-		},
-		"account_name": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Account Name",
-		},
-		"org_id": {
-			Type:        schema.TypeInt,
-			Computed:    true,
-			Description: "Customer organization identifier",
-		},
-		"organization_name": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Customer organization name",
-		},
-		"global_org_id": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Global organization identifier",
-		},
-		"global_organization_name": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Global organization name",
-		},
-		"global_cust_id": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Global Customer organization identifier",
-		},
-	}
-}
-
-func createNotificationSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"type": {
-			Type:        schema.TypeString,
-			Required:    true,
-			Description: "Notification Type - ALL,CONNECTION_APPROVAL,SALES_REP_NOTIFICATIONS, NOTIFICATIONS",
-		},
-		"send_interval": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Send interval",
-		},
-		"emails": {
-			Type:        schema.TypeList,
-			Required:    true,
-			Description: "Array of contact emails",
-			Elem: &schema.Schema{
-				Type: schema.TypeString,
-			},
-		},
-	}
-}
-
-func createErrorAdditionalInfoSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"property": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Property at which the error potentially occurred",
-		},
-		"reason": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Reason for the error",
-		},
-	}
-}
-
-func createOperationalErrorSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"error_code": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Error  code",
-		},
-		"error_message": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Error Message",
-		},
-		"correlation_id": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "CorrelationId",
-		},
-		"details": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Details",
-		},
-		"help": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Help",
-		},
-		"additional_info": {
-			Type:        schema.TypeList,
-			Computed:    true,
-			Description: "Pricing error additional Info",
-			Elem: &schema.Resource{
-				Schema: createErrorAdditionalInfoSch(),
-			},
-		},
-	}
-}
-
-var createOperationRes = &schema.Resource{
-	Schema: createOperationSch(),
-}
-
-func createOperationSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"provider_status": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Connection provider readiness status",
-		},
-		"equinix_status": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Connection status",
-		},
-		"errors": {
-			Type:        schema.TypeList,
-			Computed:    true,
-			Description: "Errors occurred",
-			Elem: &schema.Resource{
-				Schema: createOperationalErrorSch(),
-			},
-		},
-	}
-}
-
-func createAdditionalInfoSch() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"key": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Additional information key",
-		},
-		"value": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Description: "Additional information value",
-		},
-	}
-}
-
-func createFabricConnectionResourceSchema() map[string]*schema.Schema {
-	return map[string]*schema.Schema{
-		"href": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Connection URI information",
-		},
-		"name": {
-			Type:        schema.TypeString,
-			Required:    true,
-			Description: "Connection name. An alpha-numeric 24 characters string which can include only hyphens and underscores",
-		},
-		// TODO v4.ConnectionPostRequest doesn't have a description field
-		// "description": {
-		// 	Type:        schema.TypeString,
-		// 	Optional:    true,
-		// 	Description: "Customer-provided connection description",
-		// },
-		"type": {
-			Type:         schema.TypeString,
-			Required:     true,
-			ValidateFunc: validation.StringInSlice([]string{"VG_VC", "EVPL_VC", "EPL_VC", "EC_VC", "IP_VC", "IPWAN_VC", "ACCESS_EPL_VC", "EVPLAN_VC", "EPLAN_VC"}, true),
-			Description:  "Defines the connection type like VG_VC, EVPL_VC, EPL_VC, EC_VC, IP_VC, IPWAN_VC,ACCESS_EPL_VC,EVPLAN_VC, EPLAN_VC",
-		},
-		"bandwidth": {
-			Type:        schema.TypeInt,
-			Required:    true,
-			Description: "Connection bandwidth in Mbps",
-		},
-		"is_remote": {
-			Type:        schema.TypeBool,
-			Computed:    true,
-			Description: "Connection property derived from access point locations",
-		},
-		"state": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Connection overall state",
-		},
-		"operation": {
-			Type:        schema.TypeSet,
-			Computed:    true,
-			Description: "Connection type-specific operational data",
-			Elem: &schema.Resource{
-				Schema: createOperationSch(),
-			},
-		},
-		"notifications": {
-			Type:        schema.TypeList,
-			Required:    true,
-			Description: "Preferences for notifications on connection configuration or status changes",
-			Elem: &schema.Resource{
-				Schema: createNotificationSch(),
-			},
-		},
-		"additional_info": {
-			Type:        schema.TypeList,
-			Optional:    true,
-			Description: "Connection additional information",
-			Elem: &schema.Schema{
-				Type: schema.TypeMap,
-			},
-		},
-		"order": {
-			Type:        schema.TypeSet,
-			Optional:    true,
-			Description: "Order related to this connection information",
-			MaxItems:    1,
-			Elem: &schema.Resource{
-				Schema: createOrderSch(),
-			},
-		},
-		"project": {
-			Type:        schema.TypeSet,
-			Optional:    true,
-			Description: "Project information",
-			MaxItems:    1,
-			Elem: &schema.Resource{
-				Schema: createGatewayProjectSch(),
-			},
-		},
-		"account": {
-			Type:        schema.TypeSet,
-			Computed:    true,
-			Description: "Customer account information that is associated with this connection",
-			Elem: &schema.Resource{
-				Schema: createAccountSch(),
-			},
-		},
-		"change_log": {
-			Type:        schema.TypeSet,
-			Computed:    true,
-			Description: "Captures connection lifecycle change information",
-			Elem: &schema.Resource{
-				Schema: createChangeLogSch(),
-			},
-		},
-		"redundancy": {
-			Type:        schema.TypeSet,
-			Optional:    true,
-			Description: "Redundancy Information",
-			MaxItems:    1,
-			Elem: &schema.Resource{
-				Schema: createRedundancySch(),
-			},
-		},
-		"direction": {
-			Type:        schema.TypeString,
-			Computed:    true,
-			Description: "Connection directionality from the requester point of view",
-		},
-		"a_side": {
-			Type:        schema.TypeSet,
-			Required:    true,
-			Description: "Requester or Customer side connection configuration object of the multi-segment connection",
-			MaxItems:    1,
-			Elem:        createFabricConnectionSideRes(),
-			Set:         schema.HashResource(createConnectionSideAccessPointRes()),
-		},
-		"z_side": {
-			Type:        schema.TypeSet,
-			Required:    true,
-			Description: "Destination or Provider side connection configuration object of the multi-segment connection",
-			MaxItems:    1,
-			Elem:        createFabricConnectionSideRes(),
-			Set:         schema.HashResource(createConnectionSideAccessPointRes()),
-		},
-	}
-}
